refactor import from files into reusable function
